+++ backend/src/test/java/tcp/project/agenda/agenda/application/AgendaServiceTest.java
@@ -154,101 +154,6 @@ class AgendaServiceTest extends ApplicationServiceTest {
     }
 
     @Test
-    @DisplayName("투표가 되어야 함")
-    void voteTest() throws Exception {
-        //given
-        agendaService.createAgenda(regular.getId(), getBasicAgendaCreateRequest());
-
-        //when
-        agendaService.vote(regular.getId(), 1L, getBasicVoteRequest());
-
-        //then
-        List<Vote> voteList = voteRepository.findAll();
-        assertThat(voteList).hasSize(2);
-    }
-
-    @Test
-    @DisplayName("잘못된 body일 경우 예외가 발생해야 함")
-    void voteTest_invalidRequest() throws Exception {
-        //given
-
-        //when then
-        assertThatThrownBy(() -> agendaService.vote(regular.getId(), 1L, getInvalidVoteRequest()))
-                .isInstanceOf(ValidationException.class);
-    }
-
-    @Test
-    @DisplayName("없는 안건일 경우 예외가 발생해야 함")
-    void voteTest_agendaNotFound() throws Exception {
-        //given
-        Long notExistAgendaId = 999L;
-        agendaService.createAgenda(regular.getId(), getBasicAgendaCreateRequest());
-
-
-        //when then
-        assertThatThrownBy(() -> agendaService.vote(regular.getId(), notExistAgendaId, getBasicVoteRequest()))
-                .isInstanceOf(AgendaNotFoundException.class);
-    }
-
-    @Test
-    @DisplayName("없는 유저일 경우 예외가 발생해야 함")
-    void voteTest_memberNotFound() throws Exception {
-        //given
-        Long notExistMember = 999L;
-        agendaService.createAgenda(regular.getId(), getBasicAgendaCreateRequest());
-
-        //when then
-        assertThatThrownBy(() -> agendaService.vote(notExistMember, 1L, getBasicVoteRequest()))
-                .isInstanceOf(MemberNotFoundException.class);
-    }
-
-    @Test
-    @DisplayName("이미 마감된 안건에 투표할 경우 예외가 발생해야 함")
-    void voteTest_alreadyClosedAgenda() throws Exception {
-        //given
-        agendaService.createAgenda(regular.getId(), getBasicAgendaCreateRequest());
-        agendaService.closeAgenda(regular.getId(), 1L);
-
-        //when then
-        assertThatThrownBy(() -> agendaService.vote(regular.getId(), 1L, getBasicVoteRequest()))
-                .isInstanceOf(AgendaAlreadyClosedException.class);
-    }
-
-    @Test
-    @DisplayName("없는 투표 항목을 투표할 경우 예외가 발생해야 함")
-    void voteTest_agendaItemNotFound() throws Exception {
-        //given
-        agendaService.createAgenda(regular.getId(), getBasicAgendaCreateRequest());
-
-        //when then
-        assertThatThrownBy(() -> agendaService.vote(regular.getId(), 1L, getNotExistSelectItemVoteRequest()))
-                .isInstanceOf(AgendaItemNotFoundException.class);
-    }
-
-    @Test
-    @DisplayName("대상이 아닌 안건에 투표할 경우 예외가 발생해야 함")
-    void voteTest_NotTargetMember() throws Exception {
-        //given
-        agendaService.createAgenda(regular.getId(), getBasicAgendaCreateRequest());
-
-        //when then
-        assertThatThrownBy(() -> agendaService.vote(general.getId(), 1L, getNotExistSelectItemVoteRequest()))
-                .isInstanceOf(NotTargetMemberException.class);
-    }
-
-    @Test
-    @DisplayName("이미 투표한 안건에 또 투표를 하면 예외가 발생해야 함")
-    void voteTest_alreadyVote() throws Exception {
-        //given
-        agendaService.createAgenda(regular.getId(), getBasicAgendaCreateRequest());
-        agendaService.vote(regular.getId(), 1L, getBasicVoteRequest());
-
-        //when then
-        assertThatThrownBy(() -> agendaService.vote(regular.getId(), 1L, getBasicVoteRequest()))
-                .isInstanceOf(AlreadyVoteException.class);
-    }
-
-    @Test
     @DisplayName("투표가 지워져야 함")
     void voteCancelTest() throws Exception {
         //given
